refactor: Moved and renamed createForLanguageAndCurrency to SearchEngine class
