+++ src/Core/SearchEngine.php
@@ -9,7 +9,11 @@
 namespace PrestaShop\Module\Doofinder\Core;
 
 use PrestaShop\Module\Doofinder\Api\DoofinderLayerApi;
+use PrestaShop\Module\Doofinder\Api\EasyREST;
+use PrestaShop\Module\Doofinder\Configuration\DoofinderConfig;
 use PrestaShop\Module\Doofinder\Manager\LanguageManager;
+use PrestaShop\Module\Doofinder\Manager\UrlManager;
+use PrestaShop\Module\Doofinder\Utils\DfTools;
 
 if (!defined('_PS_VERSION_')) {
     exit;
@@ -88,5 +92,84 @@ class SearchEngine
         }
 
         return true;
+    }
+
+    /**
+     * Creates the Search Engine for a single language/currency combination on
+     * a single shop, via the universal install endpoint, without resending
+     * (and reprocessing) the whole store.
+     *
+     * @param int $shopId
+     * @param int $idLang
+     * @param int $idCurrency
+     *
+     * @return string|null The new Search Engine hashid, or null on failure.
+     */
+    public static function createForLanguageAndCurrency($shopId, $idLang, $idCurrency)
+    {
+        $shop = \Shop::getShop($shopId);
+        $shopGroupId = $shop['id_shop_group'];
+        $installationID = \Configuration::get('DF_INSTALLATION_ID', null, $shopGroupId, $shopId);
+
+        if (empty($installationID)) {
+            DoofinderConfig::debug("Cannot create Search Engine: shop {$shopId} has no Doofinder installation yet.");
+
+            return null;
+        }
+
+        $lang = new \Language($idLang);
+        $currency = new \Currency($idCurrency);
+
+        if (!\Validate::isLoadedObject($lang) || !\Validate::isLoadedObject($currency)) {
+            DoofinderConfig::debug("Cannot create Search Engine: invalid language ({$idLang}) or currency ({$idCurrency}).");
+
+            return null;
+        }
+
+        $existingHash = self::getHashId($idLang, $idCurrency, $shopGroupId, $shopId);
+        if ($existingHash) {
+            return $existingHash;
+        }
+
+        $multipriceEnabled = (bool) \Configuration::get('DF_MULTIPRICE_ENABLED', null, null, null, true);
+        $feedUrl = $multipriceEnabled
+            ? UrlManager::getFeedUrl($shopId, $lang->iso_code)
+            : UrlManager::getFeedUrl($shopId, $lang->iso_code, $currency->iso_code);
+
+        $client = new EasyREST();
+        $apiKey = DfTools::getFormattedApiKey();
+
+        $payload = json_encode([
+            'store_id' => $installationID,
+            'language' => $lang->language_code,
+            'currency' => $currency->iso_code,
+            'feed_url' => $feedUrl,
+            'callback_url' => UrlManager::getProcessCallbackUrl($shopId),
+        ]);
+
+        DoofinderConfig::debug("Creating Search Engine for language {$lang->language_code} / currency {$currency->iso_code} on shop {$shopId}");
+
+        $response = $client->post(
+            UrlManager::getCreateSearchEngineUrl(\Configuration::get('DF_REGION')),
+            $payload,
+            null,
+            null,
+            'application/json',
+            ['Authorization: Token ' . $apiKey]
+        );
+
+        if ($response->getResponseCode() !== 201) {
+            $message = "Failed to create Search Engine for language {$lang->language_code} / currency {$currency->iso_code} on shop {$shopId}. Response: " . $response->response;
+            DoofinderConfig::debug($message);
+            \PrestaShopLogger::addLog($message, 3, null, 'Module', null, true);
+
+            return null;
+        }
+
+        $data = json_decode($response->response, true);
+        $hashidKey = 'DF_HASHID_' . strtoupper($currency->iso_code) . '_' . strtoupper($lang->language_code);
+        \Configuration::updateValue($hashidKey, $data['hashid'], false, $shopGroupId, $shopId);
+
+        return $data['hashid'];
     }
 }
