+++ src/Installer/DoofinderInstallation.php
@@ -337,78 +337,6 @@ class DoofinderInstallation
     }
 
     /**
-     * Creates the Search Engine for a single language/currency combination on
-     * a single shop, via the universal install endpoint, without resending
-     * (and reprocessing) the whole store.
-     *
-     * @param int $shopId
-     * @param int $idLang
-     * @param int $idCurrency
-     *
-     * @return string|null The new Search Engine hashid, or null on failure.
-     */
-    public static function createSearchEngineForLanguageAndCurrency($shopId, $idLang, $idCurrency)
-    {
-        $shop = \Shop::getShop($shopId);
-        $shopGroupId = $shop['id_shop_group'];
-        $installationID = \Configuration::get('DF_INSTALLATION_ID', null, $shopGroupId, $shopId);
-
-        if (empty($installationID)) {
-            DoofinderConfig::debug("Cannot create Search Engine: shop {$shopId} has no Doofinder installation yet.");
-
-            return null;
-        }
-
-        $lang = new \Language($idLang);
-        $currency = new \Currency($idCurrency);
-
-        if (!\Validate::isLoadedObject($lang) || !\Validate::isLoadedObject($currency)) {
-            DoofinderConfig::debug("Cannot create Search Engine: invalid language ({$idLang}) or currency ({$idCurrency}).");
-
-            return null;
-        }
-
-        $multipriceEnabled = (bool) \Configuration::get('DF_MULTIPRICE_ENABLED', null, null, null, true);
-        $feedUrl = $multipriceEnabled
-            ? UrlManager::getFeedUrl($shopId, $lang->iso_code)
-            : UrlManager::getFeedUrl($shopId, $lang->iso_code, $currency->iso_code);
-
-        $client = new EasyREST();
-        $apiKey = DfTools::getFormattedApiKey();
-
-        $payload = json_encode([
-            'store_id' => $installationID,
-            'language' => $lang->language_code,
-            'currency' => $currency->iso_code,
-            'feed_url' => $feedUrl,
-            'callback_url' => UrlManager::getProcessCallbackUrl($shopId),
-        ]);
-
-        DoofinderConfig::debug("Creating Search Engine for language {$lang->language_code} / currency {$currency->iso_code} on shop {$shopId}");
-
-        $response = $client->post(
-            UrlManager::getCreateSearchEngineUrl(\Configuration::get('DF_REGION')),
-            $payload,
-            null,
-            null,
-            'application/json',
-            ['Authorization: Token ' . $apiKey]
-        );
-
-        if ($response->getResponseCode() !== 201) {
-            DoofinderConfig::debug("Failed to create Search Engine for language {$lang->language_code} / currency {$currency->iso_code} on shop {$shopId}. Response: " . $response->response);
-
-            return null;
-        }
-
-        $data = json_decode($response->response, true);
-        $hashidKey = 'DF_HASHID_' . strtoupper($currency->iso_code) . '_' . strtoupper($lang->language_code);
-        \Configuration::updateValue($hashidKey, $data['hashid'], false, $shopGroupId, $shopId);
-
-        return $data['hashid'];
-    }
-
-    /**
      * Sends a request to the plugins API to update the store options (url, shop_id, shop_group_id)
      * for every shop that has a Doofinder installation registered.
      *
